requirements.txt
openai>=1.82.0
anthropic>=0.52.0
python-dotenv>=1.0.0
fastapi
uvicorn
pyperclip
pandas>=2.2.3
